handel update image action

+++ app/Actions/CategoryActions/UpdateCategoryAction.php
@@ -11,6 +11,7 @@ class UpdateCategoryAction
 {
     public function execute(Category $category, $categoryData)
     {
+//        dd($category,$categoryData);
         $old_path = $category->image;
         $data = $categoryData;
         if (isset($categoryData['image'])) {
@@ -18,7 +19,7 @@ class UpdateCategoryAction
             $addFileAction = new StoreImageAction();
             $data['image'] = $addFileAction->execute($categoryData['image'],StoringPath::CATEGORY->value);
         }
-        if (isset($category->image)) {
+        if (isset($category->image)&& isset($data['image'])) {
             \Storage::disk('public')->delete($old_path);
         }
         $category->update($data);
